renaming tests files from example placeholder
