+++ src/pca_report_library/__main__.py
@@ -1,5 +1,5 @@
 """Code to run if this package is used as a Python module."""
 
-from .example import main
+from .report import main
 
 main()
